Refactor StarlingCurve and Terminal classes; enhance tests for CirculationModel, Motor, PressureSensor, and StarlingCurve
- Updated StarlingCurve to provide more realistic default RV and LV points.
- Improved documentation in StarlingCurve.h for clarity on functionality.
- Refactored Terminal.cpp for better readability and added enhanced output formatting.
- Removed redundant code and improved terminal I/O handling.
- Added unit tests for Motor and PressureSensor classes to ensure functionality.
- Enhanced existing tests for CirculationModel and StarlingCurve to cover more scenarios.
- Removed obsolete test_circulation file and added new test files for Motor and PressureSensor.

+++ src/model/Motor.cpp
@@ -2,71 +2,56 @@
 #include <algorithm>
 #include <cmath>
 
-Motor::Motor(int maxPressure)
+Motor::Motor(float maxPressure)
     : m_maxPressure(maxPressure)
-    , m_setpointRPM(0)
-    , m_currentRPM(0)
-    , m_currentVoltage(0) {}
+    , m_setpointRPM(0.0f)
+    , m_currentRPM(0.0f)
+    , m_currentVoltage(0.0f) {}
 
-void Motor::initialize(int defaultRPM) {
+void Motor::initialize(float defaultRPM) {
     m_currentRPM = defaultRPM;
     m_setpointRPM = defaultRPM;
     m_currentVoltage = rpmToVoltage(m_currentRPM);
 }
 
-float Motor::pressureToRPM(int pressure) const {
+float Motor::pressureToRPM(float pressure) const {
     float normalizedPressure = pressure / m_maxPressure;
-    float rpm = std::sqrt(normalizedPressure) * MAX_RPM;
-    return std::max(0.0f, std::min(MAX_RPM, rpm));
+    float rpm = std::sqrt(normalizedPressure) * static_cast<float>(MAX_RPM);
+    return std::max(0.0f, std::min(static_cast<float>(MAX_RPM), rpm));
 }
 
 float Motor::rpmToPressure(float rpm) const {
-    float normalizedRPM = rpm / MAX_RPM;
+    float normalizedRPM = rpm / static_cast<float>(MAX_RPM);
     float pressure = m_maxPressure * normalizedRPM * normalizedRPM;
     return std::min(m_maxPressure, pressure);
 }
 
 float Motor::rpmToVoltage(float rpm) const {
-    return (rpm / MAX_RPM) * MAX_VOLTAGE;
+    return (rpm / static_cast<float>(MAX_RPM)) * static_cast<float>(MAX_VOLTAGE);
 }
 
 void Motor::setTargetPressure(float targetPressure) {
     m_setpointRPM = pressureToRPM(targetPressure);
 }
 
-float Motor::getSetpointRPM() const { 
-    return m_setpointRPM; 
-}
-
-float Motor::getSetpointVoltage() const { 
-    return rpmToVoltage(m_setpointRPM); 
-}
+float Motor::getSetpointRPM() const { return m_setpointRPM; }
+float Motor::getSetpointVoltage() const { return rpmToVoltage(m_setpointRPM); }
 
 void Motor::update(float dt) {
     float diff = m_setpointRPM - m_currentRPM;
     float gain = dt / TAU;
     m_currentRPM += diff * gain;
-    m_currentRPM = std::max(0.0f, std::min(MAX_RPM, m_currentRPM));
+    m_currentRPM = std::max(0.0f, std::min(static_cast<float>(MAX_RPM), m_currentRPM));
     m_currentVoltage = rpmToVoltage(m_currentRPM);
 }
 
-float Motor::getActualPressure() const { 
-    return rpmToPressure(m_currentRPM); 
-}
-
-float Motor::getActualRPM() const { 
-    return m_currentRPM; 
-}
-
-float Motor::getActualVoltage() const { 
-    return m_currentVoltage; 
-}
+float Motor::getActualPressure() const { return rpmToPressure(m_currentRPM); }
+float Motor::getActualRPM() const { return m_currentRPM; }
+float Motor::getActualVoltage() const { return m_currentVoltage; }
 
 void Motor::reduceSpeed(float reductionFactor) {
     m_setpointRPM *= reductionFactor;
     m_currentRPM *= reductionFactor;
 }
 
-float Motor::getMaxPressure() const { 
-    return m_maxPressure; 
-}
+float Motor::getMaxPressure() const { return m_maxPressure; }
